Add Controller, simple views, and fix url paths in navbar
Add the Schedules controller. We also have an Event folder. Will discuss if we are going to keep it as Schedules or Events

Add Index view for Schedules controller.

Fix the url paths in the navbar so that we can see the basic view for Event, Login, Musician, Schedules, and Venue